# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "chandanmohanty11111@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/createlable": {
            "post": {
                "summary": "Create Label",
                "description": " Create Label ",
                "operationId": "f2f3b2d2e250a6576862cbee3d9ded1c",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "labelname"
                                ],
                                "properties": {
                                    "labelname": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label added Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/addlabelbynoteid": {
            "post": {
                "summary": "Add Label By Note Id",
                "description": " Add Label By Note Id ",
                "operationId": "e76856ae401b09d73bac8a0e4850d81a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "label_id",
                                    "note_id"
                                ],
                                "properties": {
                                    "label_id": {
                                        "type": "integer"
                                    },
                                    "note_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label Added to Note Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/displaylable": {
            "get": {
                "summary": "Display Label",
                "description": " Display Label ",
                "operationId": "4b83618bee12caf43ae85e44dce291ff",
                "requestBody": {},
                "responses": {
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/updatelable": {
            "post": {
                "summary": "Update Label",
                "description": " Update label ",
                "operationId": "98da34a4bcc7273e8b2428fdf6882d7d",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "labelname"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "labelname": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label updated Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/deletelable": {
            "post": {
                "summary": "Delete Label",
                "description": " Delete label ",
                "operationId": "f00449fdc566a0064aa2a63109627995",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label deleted Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}